requirements.txt
pandas
numpy
mysql-connector-python
python-dotenv
sqlalchemy
matplotlib
seaborn
scikit-learn
category_encoders
xgboost
-e .